Fixed some more cppcheck messages: Reduced scope of several variables Fixed compilation error on GCC 4.7
git-svn-id: https://svn.code.sf.net/p/extremetuxracer/code/trunk@278 0420edf4-82e4-42fc-9478-35b55e6d67a3

+++ textures.cpp
@@ -364,7 +364,7 @@ void CTexture::LoadTextureList () {
 			string line = list.Line (i);
 			string name = SPStrN (line, "name", "");
 			int id = SPIntN (line, "id", -1);
-			CommonTex.resize(max(CommonTex.size(), id+1));
+			CommonTex.resize(max(CommonTex.size(), (size_t)id+1));
 			string texfile = SPStrN (line, "file", "");
 			bool rep = SPIntN (line, "repeat", 0) != 0;
 			if (id >= 0) {
